move set_index_dtype to utils
