+++ audformat/core/table.py
@@ -13,7 +13,6 @@ from audformat.core.column import Column
 from audformat.core.common import (
     HeaderBase,
     HeaderDict,
-    set_index_dtype,
     to_audformat_dtype,
     to_pandas_dtype,
 )
@@ -510,7 +509,7 @@ class Base(HeaderBase):
                 level: dtype for level, dtype in dtypes.items()
                 if level in converters
             }
-            df.index = set_index_dtype(df.index, converter_dtypes)
+            df.index = utils.set_index_dtype(df.index, converter_dtypes)
             # fix columns
             for column_id in columns:
                 if column_id in converters:
@@ -690,7 +689,7 @@ class MiscTable(Base):
                 level: 'Int64' for level, dtype in zip(levels, dtypes)
                 if pd.api.types.is_integer_dtype(dtype)
             }
-            index = set_index_dtype(index, int_dtypes)
+            index = utils.set_index_dtype(index, int_dtypes)
 
             if not all(levels) or len(levels) > len(set(levels)):
                 raise ValueError(
